Adicionando logica para api-transactions chamar api-bank para creditar e debitar saldos em conta

+++ Transactions/src/main/java/com/Transactions/Transactions/Dto/TransactionsCreationDto.java
@@ -9,7 +9,7 @@ import com.Transactions.Transactions.Enum.Type;
 
 public record TransactionsCreationDto(
         Type type,
-        Double amount,
+        BigDecimal amount,
         int bankAccountId,
         String counterpartyBankCode,
         String counterpartyBankName,
